+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- [runtime] `fromfile` Records applied configuration (from offspot.yaml) to /etc/offspot/latest.yaml
+- [runtime] `ap` now supports `profile` (`perf` or `coverage`)
+- [builder] Admin UI support via `.add_adminui()`
+
 ## [2.7.1] - 2025-10-20
 
 ### Changed